I made a mistake
I made a mistake in making my branches in the same file, so trying to
merge them together was throwing a bunch of errors. Thankfully I caught
it early so I was able to just copy the data I needed from the branches
and manually change master. I will now delete the branches once I know
things are saved in master